+++ kb/kb0001.md
@@ -1,51 +1,40 @@
 # HOWTO: Use Word macros to switch font and keyboard at the same time
 
-<p>To create a macro using the Keyman API that switches font and keyboard at the same time:</p>
+## Steps
 
-<ol>
-<li>Download the "Keyman API (VBA Module)” archive from the Tavultesoft Keyman Downloads.</li>
-<li>Open this zip file and extract the file KeymanLink.bas to your hard drive.</li>
-<li>Start Word, and press Alt+F11 to open the Visual Basic Editor</li>
-<li>Right-click on “Normal” and select “Import File”, and choose KeymanLink.bas where you saved it on your hard disk.  (After this you can delete KeymanLink.bas again as a copy will be stored in Word separately) </li>
-<li>Right-click on “Normal” in the Project window, and select Insert->Module </li>
-<li>Copy and paste the following code, changing the keyboard and font names as appropriate: 
- 
-<blockquote><pre>
-Sub ActivateGreek()
+To create a macro using the Keyman API that switches font and keyboard at the same time:
 
-    SelectKeymanKeyboard "Greek"
+1.  Download [Keyman API (VBA Module)](assets/kb0001/keymanlink.zip) zip archive.
+2.  Open this zip file and extract the file KeymanLink.bas to your hard drive.
+3.  Start Word, and press Alt+F11 to open the Visual Basic Editor
+4.  Right-click on “Normal” and select “Import File”, and choose KeymanLink.bas where you saved it on your hard disk. (After this you can delete KeymanLink.bas again as a copy will be stored in Word separately)
+5.  Right-click on “Normal” in the Project window, and select Insert->Module
+6.  Copy and paste the following code, changing the keyboard and font names as appropriate:
 
-    Selection.Font.Name = "MyGreekFont"
+<!-- css gives reasonable approximation of vb here -->
 
-End Sub
+```css
+    Sub ActivateGreek()
+        SelectKeymanKeyboard "Greek"
+        Selection.Font.Name = "MyGreekFont"
+    End Sub
 
- 
+    Sub ActivateHebrew()
+        SelectKeymanKeyboard "Hebrew"
+        Selection.Font.Name = "MyHebrewFont"
+    End Sub
 
-Sub ActivateHebrew()
+    Sub ActivateEnglish()
+        SelectKeymanKeyboard "", False
+        Selection.Font.Name = "Times New Roman"
+    End Sub
+```
 
-    SelectKeymanKeyboard "Hebrew"
-
-    Selection.Font.Name = "MyHebrewFont"
-
-End Sub
-
- 
-
-Sub ActivateEnglish()
-
-    SelectKeymanKeyboard "", False
-
-    Selection.Font.Name = "Times New Roman"
-
-End Sub
-</pre></blockquote></li>
-
-<li>Close the Visual Basic window </li>
-<li>Select Tools->Options->Customise, and click the Keyboard button. </li>
-<li>Scroll down the list to Macros, and you will see your macros on the right hand list. </li>
-<li>Select each macro in turn, click in the “Press new Shortcut Key” window, and press the hotkey to run that macro, then click Assign to use that hotkey.  You will want to be careful not to override a hotkey you use normally for other actions. </li>
-<li>Select Close, Close again.  When you close Word, you will be prompted to save “normal.dot”.  Make sure you answer Yes to save your macros. </li>
-</ol>
+7.  Close the Visual Basic window
+8.  Select Tools->Options->Customise, and click the Keyboard button.
+9.  Scroll down the list to Macros, and you will see your macros on the right hand list.
+10.  Select each macro in turn, click in the “Press new Shortcut Key” window, and press the hotkey to run that macro, then click Assign to use that hotkey. You will want to be careful not to override a hotkey you use normally for other actions.
+11.  Select Close, Close again. When you close Word, you will be prompted to save “normal.dot”. Make sure you answer Yes to save your macros.
 
 ## Applies to:
  * Keyman 6.0
