+++ kb/kb0002.md
@@ -1,28 +1,28 @@
 # HOWTO: Report errors with Keyman
 
-<h1>Reporting Errors and Problems</h1><p>
-If you are having difficulty with using a particular keyboard layout, please contact the developer of that keyboard first.
-</p><p>
-If the problem is with installing or using Keyman, before emailing Tavultesoft, please check if any Knowledge Base articles
-(see <a href="http://www.tavultesoft.com/kb/">http://www.tavultesoft.com/kb/</a>)
-may be relevant.
+## Reporting Errors and Problems
 
-</p><p>
-When contacting Tavultesoft, as well as advising the version (and release) of Keyman 
-you are using (e.g. 6.2.178.9) and keyboard layout files, please indicate the operating 
+If you are having difficulty with using a particular keyboard layout, please contact the developer of that keyboard first.
+
+If the problem is with installing or using Keyman, before emailing Tavultesoft, please check if any Knowledge Base articles
+(see <a href="http://www.tavultesoft.com/kb/">http://www.tavultesoft.com/kb/</a>) may be relevant.
+
+When contacting Tavultesoft, as well as advising the version (and release) of Keyman
+you are using (e.g. 6.2.178.9) and keyboard layout files, please indicate the operating
 system version (Windows 95/98/Me/NT/2000/XP etc), total memory (RAM), and whether
 other language packs are installed (e.g. Chinese/Japanese/Korean IME).
-</p><p>The Standard Edition and Corporate Editions of Keyman include a Diagnostic Report, available through Keyman Configuration, Support tab, that will generate a comprehensive report on the state of your system, and let you send that to Tavultesoft for review.
-</p><p>
+
+The Standard Edition and Corporate Editions of Keyman include a Diagnostic Report, available through Keyman Configuration, Support tab, that will generate a comprehensive report on the state of your system, and let you send that to Tavultesoft for review.
+
 Please provide a detailed description of the problem you are having, and if Keyman is crashing, it would be helpful if you can use Dr Watson (drwtsn32.exe on Windows 2000 or XP)
 to record detailed logs of the fault.  Record which programs you are using, and exactly what steps you took that caused the problem.  If the problem is intermittent, try and identify when the problem was most likely to happen.
-</p><p>
-If you are having problems with a keyboard that you have written, include portions or the whole of your keyboard source file (.kmn) for us to test.  These source files will 
+
+If you are having problems with a keyboard that you have written, include portions or the whole of your keyboard source file (.kmn) for us to test.  These source files will
 be kept in confidence and destroyed when we have closed your support issue.
-</p>
-<h2>Note!</h2><p>
+
+## Note!
+
 When reporting errors, be sure to provide your Keyman licence details, as licenced users get priority in our replies.
-</p>
 
 ## Applies to:
  * Keyman 5.0
