+++ kb/kb0094.md
@@ -44,3 +44,5 @@
 
 ## Applies to:
  * Keyman Desktop 9.0
+ * Keyman Desktop 10.0
+ * Keyman Desktop 11.0
